+++ WPFWijkagent/WijkagentModels/Location.cs
@@ -10,7 +10,7 @@ namespace WijkagentModels
         public double Longitude { get; set; }
         public double Latitude { get; set; }
 
-        public Location( double longtitude, double latitude)
+        public Location( double latitude, double longtitude)
         {
             Longitude = longtitude;
             Latitude = latitude;
