+++ WPFWijkagent/WijkagentTests/ScraperTest.cs
@@ -46,18 +46,16 @@ namespace WijkagentTests
             //arrange
             Offence offence = new Offence()
             {
-                ID = 3,
-                DateTime = new DateTime().ToLocalTime(),
-                Description = "een delict..",
-                LocationID = new Location(52.501127, 6.0789937)
+                LocationID = new Location(52.501127, 6.0789937),
+                DateTime = new DateTime(2019, 11, 25)
             };
-            Scraper scraper = new Scraper(offence);
+            Scraper s = new Scraper(offence);
             //act
-            Scraper.Connect();
-            List<SocialMediaMessage> result = scraper.GetSocialMediaMessages();
+            List<SocialMediaMessage> socials = s.GetSocialMediaMessages();
+            Console.WriteLine(socials.Count);
             //assert
-            Assert.IsNotNull(result);
-            Assert.GreaterOrEqual(result.Count, 1);
+            Assert.Greater(socials.Count, 0);
+
         }
     }
 }
