+++ WPFWijkagent/WijkagentTests/SocialMediaDialogueTest.cs
@@ -1,14 +1,74 @@
-﻿using NUnit.Framework;
+﻿using Microsoft.Maps.MapControl.WPF;
+using NUnit.Framework;
 using System;
 using System.Collections.Generic;
 using System.Text;
+using WijkagentModels;
+using WijkagentWPF;
+using Location = WijkagentModels.Location;
 
 namespace WijkagentTests
 {
     [TestFixture]
     class SocialMediaDialogueTest
     {
-        public void 
+        //private Offence offence_1 = new Offence()
+        //{
+        //    ID = 3,
+        //    DateTime = new DateTime().ToLocalTime(),
+        //    Description = "een delict..",
+        //    LocationID = new Location(52.501127, 6.0789937)
+        //};
 
+        //private Offence offence_2 = new Offence()
+        //{
+        //    ID = 3,
+        //    DateTime = new DateTime().ToLocalTime(),
+        //    Description = "een delict..",
+        //    LocationID = new Location(53.504127, 6.0789437)
+        //};
+
+        //private Pushpin Pins = new Pushpin()
+        //{
+        //    Location = new Microsoft.Maps.MapControl.WPF.Location
+        //    {
+        //        Latitude = 52.501127,
+        //        Longitude = 6.0789937
+        //    },
+        //};
+
+        //[Test]
+        //public void RetrieveOffence_OffenceGiven_ReturnsCorrectPin()
+        //{
+        //    //arrange
+        //    List<OffenceListItem> items = new List<OffenceListItem>() {new OffenceListItem(offence_1), new OffenceListItem(offence_2)}; 
+        //    SocialMediaDialogueController sd = new SocialMediaDialogueController(Pins,items);
+        //    //act
+        //    Offence o = sd.RetrieveOffence();
+        //    //Assert
+        //    Assert.AreEqual(offence_1, o);
+        //}
+
+        //[Test]
+        //public void DisplayMessages_OffencesFound_ReturnsString()
+        //{
+        //    //arrange
+        //    List<OffenceListItem> items = new List<OffenceListItem>() { new OffenceListItem(offence_1), new OffenceListItem(offence_2) };
+        //    SocialMediaDialogueController sd = new SocialMediaDialogueController(Pins, items);
+        //    //act
+        //    string result = ;
+        //    //Assert
+        //    Assert.GreaterOrEqual(0, result.Length);
+        //}
+
+        //[Test]
+        //public void MethodName_SituationGiven_ExpectedReturn()
+        //{
+        //    //arrange
+
+        //    //act
+
+        //    //Assert
+        //}
     }
 }
